Extract the common code between the 2 Day6 part in a common local library.
