+++ 2015/Day6/CMakeLists.txt
@@ -1,12 +1,16 @@
 cmake_minimum_required(VERSION 3.12.4)
 
+add_library(ProbablyAFireHazard_SharedCode src/rectangle.cpp src/input_parser.cpp )
+set_target_properties(ProbablyAFireHazard_SharedCode PROPERTIES CXX_STANDARD 17)
+target_link_libraries(ProbablyAFireHazard_SharedCode PUBLIC coordinate)
+
 add_executable(ProbablyAFireHazard_Part1 src/part_one.cpp)
 set_target_properties(ProbablyAFireHazard_Part1 PROPERTIES CXX_STANDARD 17)
-target_link_libraries(ProbablyAFireHazard_Part1 PUBLIC stringManipulation coordinate)
+target_link_libraries(ProbablyAFireHazard_Part1 PUBLIC stringManipulation ProbablyAFireHazard_SharedCode)
 
 add_executable(ProbablyAFireHazard_Part2 src/part_two.cpp)
 set_target_properties(ProbablyAFireHazard_Part2 PROPERTIES CXX_STANDARD 17)
-target_link_libraries(ProbablyAFireHazard_Part2 PUBLIC stringManipulation coordinate)
+target_link_libraries(ProbablyAFireHazard_Part2 PUBLIC stringManipulation ProbablyAFireHazard_SharedCode)
 target_link_options(ProbablyAFireHazard_Part2 PRIVATE /STACK:10000000)
 
 add_test(NAME ProbablyAFireHazard_part_one_puzzle_input
