+++ 2015/Day6/src/input_parser.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "light_action.hpp"
+#include "rectangle.hpp"
+#include <string_view>
+
+LightAction getActionFromLine(const std::string_view line);
+
+Rectangle getRectangleFromLine(std::string line);
