+++ 2015/Day6/src/light_action.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+enum class LightAction
+{
+    TURN_ON,
+    TURN_OFF,
+    TOGGLE,
+    NONE,
+};
