+++ 2015/Day6/src/rectangle.cpp
@@ -1,0 +1,12 @@
+#include "rectangle.hpp"
+
+void foreachCoordinateInRectangle(Rectangle rect, std::function<void(const Coordinate)> callback)
+{
+        for(auto x = rect.topLeft.x; x <= rect.bottomRight.x; ++x)
+        {
+            for(auto y = rect.topLeft.y; y <= rect.bottomRight.y; ++y)
+            {
+                callback({x, y});
+            }
+        }
+}
