+++ 2015/Day6/src/rectangle.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "coordinate.hpp"
+#include <functional>
+
+struct Rectangle
+{
+    Coordinate topLeft, bottomRight;
+};
+
+void foreachCoordinateInRectangle(Rectangle rect, std::function<void(const Coordinate)> callback);
